【Infer Symbolic Shape No.242】【BUAA】`warpctc`  (#67539)
* Update ops.yaml multiary_infer_sym.cc multiary_infer_sym.h Warpctc Op + unary_infer_sym.cc OneHot Op + binary_infer_sym.cc Unpool3d Op

* Update multiary_infer_sym.cc binary_infer_sym.cc

* Update multiary_infer_sym.cc

* Update multiary_infer_sym.cc

* Update unary_infer_sym.cc unary.cc

* Update unary_infer_sym.cc unary.cc

* Update unary_infer_sym.cc

* Update multiary.cc + multiary_infer_sym.cc

* Update multiary.cc

* Update multiary.cc

* Update multiary.cc

* Update unary_infer_sym.cc one_hot op

+++ paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/binary_infer_sym.cc
@@ -1474,7 +1474,7 @@ bool Unpool3dOpInferSymbolicShape(
       details::GetVectorAttr<int>(op, "output_size");
 
   std::vector<symbol::DimExpr> output_shape = {x_shape[0], x_shape[1]};
-  for (int i = 0; i < static_cast<int>(ksize.size()); ++i) {
+  for (size_t i = 0; i < ksize.size(); ++i) {
     output_shape.emplace_back(symbol::DimExpr(output_size[i]));
   }
